Refactoring & cleanup
